added vectorized Learner, double precision
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2094 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/Grid.cpp
@@ -295,7 +295,7 @@ void Grid::insertPoint(size_t dim, unsigned int levels[], unsigned int indices[]
 	storage->insert(pointIndex);
 }
 
-int Grid::getSize(){
+size_t Grid::getSize(){
 	return this->storage->size();
 }
